fix: [DHIS2-21266] Apply changes from PR feedback
* Folder structure
* Update cypress tests
