+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2026-06-02T09:46:52.775Z\n"
-"PO-Revision-Date: 2026-06-02T09:46:52.775Z\n"
+"POT-Creation-Date: 2026-06-17T22:59:57.403Z\n"
+"PO-Revision-Date: 2026-06-17T22:59:57.404Z\n"
 
 msgid "The application could not be loaded."
 msgstr "The application could not be loaded."
