+++ src/core_modules/capture-core/components/WidgetFeedback/WidgetFeedbackContent/MarkdownWrapper/MarkdownWrapper.types.ts
@@ -1,0 +1,5 @@
+export type MarkdownWrapperProps = {
+  title: string;
+  content?: string;
+  color?: string;
+}
