+++ src/core_modules/capture-core/components/WidgetFeedback/WidgetFeedbackContent/MarkdownWrapper/index.ts
@@ -1,0 +1,1 @@
+export { MarkdownWrapper } from './MarkdownWrapper';
